Add Makefile for mixer/v1 proto generation (#277)
This PR removes the mixer/v1 bazel code for protoc-based generation, replacing it that logic with a Makefile.

Notes:

- generate-protos.sh is updated to call make targets
- all BUILD files were rewritten with gazelle to supply default targets around the generated files
- each of the .proto files is touched, as it is regenerated as part of this PR

+++ scripts/generate-protos.sh
@@ -3,6 +3,8 @@
 set -o errexit
 set -o nounset
 
+make clean generate
+
 bazel build //...
 
 genfiles=$(bazel info bazel-genfiles)
@@ -15,4 +17,6 @@ for src in ${files}; do
     fi
 done
 
-rm mixer/v1/config/cfg.pb.go
+if [ -e mixer/v1/config/cfg.pb.go ]; then
+    rm mixer/v1/config/cfg.pb.go
+fi
